fix: change play count fields from Int to Float in GraphQL schema
Play count values exceed GraphQL's 32-bit Int limit (2,147,483,647).
Values like 3,646,673,061 caused production errors:
"Int cannot represent non 32-bit signed integer value"

Changes:
- Updated TrackPlayCountType: all *_current_play_count fields now Float
- Updated TrackType: total/spotify/youtube play count fields now Float
- Frontend already uses abbreviated strings for display, so Float precision is fine

This fixes the production GraphQL query errors preventing playlist loading.

+++ backend/gql/play_count.py
@@ -12,31 +12,31 @@ class TrackPlayCountType:
 
     isrc: str
 
-    spotify_current_play_count: int | None = None
+    spotify_current_play_count: float | None = None
     spotify_weekly_change_percentage: float | None = None
     spotify_updated_at: datetime | None = None
     spotify_current_play_count_abbreviated: str | None = None
     spotify_weekly_change_percentage_formatted: str | None = None
 
-    apple_music_current_play_count: int | None = None
+    apple_music_current_play_count: float | None = None
     apple_music_weekly_change_percentage: float | None = None
     apple_music_updated_at: datetime | None = None
     apple_music_current_play_count_abbreviated: str | None = None
     apple_music_weekly_change_percentage_formatted: str | None = None
 
-    youtube_current_play_count: int | None = None
+    youtube_current_play_count: float | None = None
     youtube_weekly_change_percentage: float | None = None
     youtube_updated_at: datetime | None = None
     youtube_current_play_count_abbreviated: str | None = None
     youtube_weekly_change_percentage_formatted: str | None = None
 
-    soundcloud_current_play_count: int | None = None
+    soundcloud_current_play_count: float | None = None
     soundcloud_weekly_change_percentage: float | None = None
     soundcloud_updated_at: datetime | None = None
     soundcloud_current_play_count_abbreviated: str | None = None
     soundcloud_weekly_change_percentage_formatted: str | None = None
 
-    total_current_play_count: int | None = None
+    total_current_play_count: float | None = None
     total_weekly_change_percentage: float | None = None
     total_current_play_count_abbreviated: str | None = None
     total_weekly_change_percentage_formatted: str | None = None
